Include y-intercept with dtdnu fits. This parameter may contain physical info
y-intercept is in units of ms so I'm calling it "t_b"
In the TRDM it is equivalent to nu0 * dtau'/dnu.
If t_b is small (<1ms) it may indicate that the burst is fully resolved.
If t_b is larger it may indicate that there are multiple sub-bursts
involved in the measurement. This is because of how t_b is derived
in the TRDM.

* add column tb (ms) for the intercept found
* add default values when fit fails
* remove measuredtdnu, include it in the main script
* update measurebursts.py with other datasets